+++ components/cascader/index.js
@@ -73,6 +73,7 @@ export default class Cascader extends Select {
     handleProps(props, labelObj, level) {
         const {_value, value} = this.get();
         let active = false;
+        let selected = false;
 
         if (Array.isArray(_value)) {
             if (_value[level] === props.value) {
@@ -83,10 +84,11 @@ export default class Cascader extends Select {
         if (Array.isArray(value)) {
             if (value[level] === props.value) {
                 labelObj.values.push(props);
+                selected = true;
             }
         }
 
-        return {active};
+        return {active, selected};
     }
 
     _onSearch(e) {