+++ components/datepicker/time.js
@@ -1,6 +1,6 @@
 import Intact from 'intact';
 import template from './time.vdt';
-import {createDate} from './utils';
+import dayjs from 'dayjs';
 
 export default class DatepickerTime extends Intact {
     @Intact.template()
@@ -9,11 +9,11 @@ export default class DatepickerTime extends Intact {
     static propTypes = {
         data: Array,
         value: Array,
-        min: [String, Date],
-        max: [String, Date],
+        min: dayjs,
+        max: dayjs,
         date: {
             required: true,
-            type: String
+            type: dayjs 
         },
         disabledItems: Array,
     };
@@ -33,11 +33,18 @@ export default class DatepickerTime extends Intact {
 
     _init() {
         this.on('$receive:value', (c, v) => {
-            this.set('_value', v);
+            this.set('_value', v, {silent: true});
         });
         this.on('$change:_value', (c, v) => {
             if (!this._isDisabled(v)) {
-                this.set('value', v);
+                this._shouldTriggerChange = true;
+                this.set('value', v, {async: true});
+                this._shouldTriggerChange = false;
+            }
+        });
+        this.on('$change:value', (c, v) => {
+            if (this._shouldTriggerChange) {
+                this.trigger('change', this, v);
             }
         });
     }
@@ -61,15 +68,9 @@ export default class DatepickerTime extends Intact {
 
         const {min, max, date} = this.get();
 
-        value = createDate(`${date} ${value.join(':')}`);
+        value = date.hour(+value[0]).minute(+value[1]).second(+value[2]);
 
-        if (min && max) {
-            return value < createDate(min) || value > createDate(max);
-        } else if (min) {
-            return value < createDate(min);
-        } else if (max) {
-            return value > createDate(max);
-        }
+        return min && value.isBefore(min) || max && value.isAfter(max);
     }
 
     _beforeUpdate(vNode) {
@@ -79,7 +80,9 @@ export default class DatepickerTime extends Intact {
         if (vNode) {
             const {_value} = this.get();
             if (!this._isDisabled(_value)) {
-                this.set('value', _value);
+                this._shouldTriggerChange = true;
+                this.set('value', _value, {async: true});
+                this._shouldTriggerChange = false;
             }
         }
     }