+++ components/form/index.js
@@ -1,7 +1,5 @@
 import Form from './form';
 import FormItem from './formItemWrapper';
-import '../../styles/kpc.styl';
-import './index.styl';
 
 export default Form;
 export {Form, FormItem}