+++ components/table/index.react.spec.js
@@ -3,12 +3,21 @@ import ReactDOM from 'react-dom';
 import {mount, unmount} from '../../test/utils';
 import {Table, TableColumn} from 'kpc/components/table';
 import {Switch} from 'kpc/components/switch';
+import {Tooltip} from 'kpc/components/tooltip';
 
 describe('Table', () => {
-    it('render component in template', () => {
-        const container = document.createElement('div');
-        document.body.appendChild(container);
+    let container;
 
+    beforeEach(() => {
+        container = document.createElement('div');
+        document.body.appendChild(container);
+    });
+
+    afterEach(() => {
+        document.body.removeChild(container);
+    });
+
+    it('render component in template', () => {
         ReactDOM.render(
             <Table data={[{}]}>
                 <TableColumn
@@ -23,7 +32,24 @@ describe('Table', () => {
             </Table>,
             container
         );
+    });
 
-        document.body.removeChild(container);
+    it('render Tooltip in template, and it does not exist first time', () => {
+        class Demo extends React.Component {
+            state = {data: []};
+            render() {
+                window.test = this;
+                return <Table data={this.state.data}>
+                    <TableColumn
+                        key="a"
+                        title="test"
+                        template={(data) => {
+                            return <Tooltip content="test">{data.a}</Tooltip>
+                        }}
+                    />
+                </Table>
+            }
+        }
+        ReactDOM.render(<Demo />, container);
     });
 });