+++ components/table/row.js
@@ -3,5 +3,18 @@ import template from './row.vdt';
 
 // for tr $destroyed event
 export default class TableRow extends Intact {
-    get template() { return template; }
+    @Intact.template()
+    static template = template;
+
+    _onClick(e) {
+        this.trigger('click', e); 
+    }
+
+    _onMouseEnter(e) {
+        this.trigger('mouseenter', e);
+    }
+
+    _onMouseLeave(e) {
+        this.trigger('mouseleave', e);
+    }
 }