+++ components/timepicker/index.spec.js
@@ -2,6 +2,7 @@ import BasicDemo from '~/components/timepicker/demos/basic';
 import RangeDemo from '~/components/timepicker/demos/range';
 import StepDemo from '~/components/timepicker/demos/step';
 import StepRangeDemo from '~/components/timepicker/demos/stepRange';
+import FormatDemo from '~/components/timepicker/demos/format';
 import {mount, unmount, dispatchEvent, getElement} from 'test/utils';
 
 describe('Timepicker', () => {
@@ -88,5 +89,35 @@ describe('Timepicker', () => {
         dropdown.querySelector('.k-disabled').nextElementSibling.click();
         expect(dropdown.innerHTML).to.matchSnapshot();
         expect(instance.get('time')).to.eql(['00:30:00', '00:30:00']);
+    });
+
+    it('format', () => {
+        instance = mount(FormatDemo);
+
+        const [input1, input2, input3] = instance.element.querySelectorAll('.k-datepicker > div');
+        input1.click();
+        let content = getElement('.k-dropdown-menu.k-time-picker');
+        expect(content.innerHTML).to.matchSnapshot();
+        // is '02'
+        content.querySelectorAll('.k-scroll-item:not(.k-disabled)')[1].click();
+        expect(instance.get('time1')).to.eql('02:00 am');
+        expect(input1.innerHTML).to.matchSnapshot();
+
+        input2.click();
+        content = getElement('.k-select-dropdown');
+        expect(content.innerHTML).to.matchSnapshot();
+        content.querySelector('.k-item').click();
+        expect(instance.get('time2')).to.eql('00:00:00');
+        expect(input2.innerHTML).to.matchSnapshot();
+
+        input3.click();
+        content = getElement('.k-dropdown-menu.k-time-picker');
+        expect(content.innerHTML).to.matchSnapshot();
+        const [scroll1, scroll2] = content.querySelectorAll('.k-scroll-select');
+        scroll1.querySelector('.k-active').nextElementSibling.click();
+        scroll2.querySelector('.k-active').nextElementSibling.nextElementSibling.click();
+        expect(content.innerHTML).to.matchSnapshot();
+        expect(instance.get('time3')).eql(['00:30:00.000', '01:00:00.000']);
+        expect(input3.innerHTML).to.matchSnapshot();
     });
 });